third_party/blink/web_tests/flag-specific/highdpi/virtual/container-queries/wpt_internal/css/css-conditional/container-queries/at-container-008-expected.txt

This is a testharness.js-based test.
FAIL Querying nearest (unnamed) container assert_equals: expected "rgb(0, 128, 0)" but got "rgb(0, 0, 0)"
FAIL Querying nearest container by name assert_equals: expected "rgb(0, 128, 0)" but got "rgb(0, 0, 0)"
FAIL Querying intermediate container by name assert_equals: expected "rgb(0, 128, 0)" but got "rgb(0, 0, 0)"
FAIL Querying topmost container by name assert_equals: expected "rgb(0, 128, 0)" but got "rgb(0, 0, 0)"
Harness: the test ran to completion.
